requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\requirements.txt'
#
aiohttp==3.6.3            # via discord.py
async-timeout==3.0.1      # via aiohttp
attrs==20.3.0             # via aiohttp
chardet==3.0.4            # via aiohttp
colorama==0.4.4           # via loguru
discord.py==1.5.1         # via -r .\requirements.txt
idna==2.10                # via yarl
loguru==0.5.3             # via -r .\requirements.txt
multidict==4.7.6          # via aiohttp, yarl
win32-setctime==1.0.3     # via loguru
yarl==1.5.1               # via aiohttp
